delete validations.go
