+++ action/candidateregister_test.go
@@ -12,6 +12,7 @@ import (
 	"testing"
 
 	"github.com/pkg/errors"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/proto"
 
@@ -145,4 +146,49 @@ func TestCandidateRegisterABIEncodeAndDecode(t *testing.T) {
 	require.Equal(test.Duration, stake.Duration())
 	require.Equal(test.AutoStake, stake.AutoStake())
 	require.Equal(test.Payload, stake.Payload())
+}
+
+func TestIsValidCandidateName(t *testing.T) {
+	tests := []struct {
+		input  string
+		output bool
+	}{
+		{
+			input:  "abc",
+			output: true,
+		},
+		{
+			input:  "123",
+			output: true,
+		},
+		{
+			input:  "abc123abc123",
+			output: true,
+		},
+		{
+			input:  "Abc123",
+			output: false,
+		},
+		{
+			input:  "Abc 123",
+			output: false,
+		},
+		{
+			input:  "Abc-123",
+			output: false,
+		},
+		{
+			input:  "abc123abc123abc123",
+			output: false,
+		},
+		{
+			input:  "",
+			output: false,
+		},
+	}
+
+	for _, tt := range tests {
+		output := IsValidCandidateName(tt.input)
+		assert.Equal(t, tt.output, output)
+	}
 }
